move transfer sheet to own package, split coordinator into own file, update coordinator to have lambda to load content from callsite/viewmodel
